+++ src/main/java/com/motivation/mojaty/domain/notification/web/dto/request/NotificationRequestDto.java
@@ -1,0 +1,10 @@
+package com.motivation.mojaty.domain.notification.web.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class NotificationRequestDto {
+
+    private String hour;
+    private String minute;
+}
